% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChisqDist.R
\name{ChisqDist}
\alias{ChisqDist}
\title{Plot a Chi-squared distribution}
\usage{
ChisqDist(df)
}
\arguments{
\item{df}{Degrees of freedom}
}
\description{
Given the degrees of freedom (\code{df}), plot a chi-squared
distribution.
}
\examples{
\dontrun{
if (require(manipulate)) {
  manipulate( chisq_dist(df), df = slider(1, 20) )
}}
}
\author{
Kevin Middleton (\email{middletonk@missouri.edu})
}

