\name{EightBall}
\alias{EightBall}
\title{EightBall}
\usage{
  EightBall(question)
}
\arguments{
  \item{question}{A string that you would like the eight
  ball to evaluate.}
}
\description{
  Ask the Magic Eight Ball
}
\details{
  Ask the magic eight ball a question, and it will reply.
}
\examples{
EightBall("Should I learn S4 methods?")
}
\author{
  Kevin Middleton
}

