% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDI_test.R
\name{HPDI_test}
\alias{HPDI_test}
\title{HPDI Interval Test}
\usage{
HPDI_test(samples, test_value = 0, upper = 0.99, lower = 0.01)
}
\arguments{
\item{samples}{1d vector of posterior samples}

\item{test_value}{Numeric value to test against}

\item{upper}{Upper bound of intervals to test. Defaults to 0.99.}

\item{lower}{Upper bound of intervals to test. Defaults to 0.99.}
}
\value{
Widest HPDI not including the test value.
}
\description{
HPDI Interval Test
}
