% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_map2stan.R
\name{unpack_map2stan}
\alias{unpack_map2stan}
\title{Unpack map or map2stan fitted model}
\usage{
unpack_map2stan(fm, ...)
}
\arguments{
\item{fm}{Fitted model}

\item{...}{Additional arguments passed on to `rethinking::extract.samples()`}
}
\value{
data.frame of fm
}
\description{
Unpack map or map2stan fitted model
}
