% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EightBall.R
\name{EightBall}
\alias{EightBall}
\title{Ask the Magic Eight Ball}
\usage{
EightBall(question)
}
\arguments{
\item{question}{A string that you would like the eight ball to
evaluate.}
}
\description{
Ask the magic eight ball a question, and it will reply.
}
\examples{
EightBall("Should I learn S4 methods?")
}
\author{
Kevin Middleton (\email{middletonk@missouri.edu})
}

