% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmmisc.R
\docType{package}
\name{kmmisc-package}
\alias{kmmisc-package}
\title{kmmisc}
\description{
Miscellaneous functions
}
\details{
\tabular{ll}{
Package: \tab kmmisc\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2020-01-25\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\c
}
}
\author{
Kevin Middleton \email{middletonk@missouri.edu}

Maintainer: Kevin Middleton \email{middletonk@missouri.edu}
}
\keyword{package}
