\docType{package}
\name{kmmisc-package}
\alias{kmmisc-package}
\title{KM Misc}
\description{
  Misc. Functions
}
\details{
  \tabular{ll}{ Package: \tab kmmisc\cr Type: \tab
  Package\cr Version: \tab 0.1-1\cr Date: \tab
  2011-08-13\cr License: \tab GPL-2\cr LazyLoad: \tab
  yes\cr LazyData: \tab yes\cr }
}
\author{
  Kevin Middleton (kmm@csusb.edu)

  Maintainer: Kevin Middleton (kmm@csusb.edu)
}
\keyword{package}

