\docType{package}
\name{kmmisc-package}
\alias{kmmisc-package}
\title{kmmisc}
\description{
  Miscellaneous functions
}
\details{
  \tabular{ll}{ Package: \tab kmmisc\cr Type: \tab
  Package\cr Version: \tab 0.1-3\cr Date: \tab
  2012-09-06\cr License: \tab GPL-2\cr LazyLoad: \tab
  yes\cr LazyData: \tab yes\cr }
}
\author{
  Kevin Middleton \email{middletonk@missouri.edu}

  Maintainer: Kevin Middleton
  \email{middletonk@missouri.edu}
}
\keyword{package}

